Add files via upload
I created a testing server using node.js and started trying to test the website there. I encountered an error where the webpage could not load until a favicon had been made and the website could display it. I made a placeholder favicon and tried my best to get the website to display it and continue testing but I was unable. I will need to do more research on what the problem could be later and hopefully make it so the website is able to run on the server before next week.